+++ packages/shared/src/task/sync.ts
@@ -2,11 +2,29 @@ import { parse } from "node:path";
 
 import { appConfig } from "../config.js";
 import { SyncTask, taskQueue } from "./task.js";
-import { BaiduPCS } from "../sync/index.js";
+import { BaiduPCS, AliyunPan } from "../sync/index.js";
 
-const getConfig = () => {
+import type { SyncType } from "@biliLive-tools/types";
+
+const getConfig = (type: SyncType) => {
   const config = appConfig.getAll();
-  return { binary: config.sync.baiduPCS.execPath, target: config.sync.baiduPCS.targetPath };
+  return { binary: config.sync[type].execPath, target: config.sync[type].targetPath };
+};
+
+const createUploadInstance = (type: SyncType, execPath: string, remotePath: string) => {
+  if (type === "baiduPCS") {
+    return new BaiduPCS({
+      binary: execPath,
+      remotePath: remotePath ?? "",
+    });
+  } else if (type === "aliyunpan") {
+    return new AliyunPan({
+      binary: execPath,
+      remotePath: remotePath ?? "",
+    });
+  } else {
+    throw new Error("Unsupported type");
+  }
 };
 
 export const addSyncTask = async ({
@@ -15,18 +33,17 @@ export const addSyncTask = async ({
   execPath,
   retry,
   policy,
+  type,
 }: {
   input: string;
   remotePath?: string;
   execPath?: string;
   retry?: number;
   policy?: "fail" | "newcopy" | "overwrite" | "skip" | "rsync";
+  type: SyncType;
 }) => {
-  const { binary: binaryPath, target: targetPath } = getConfig();
-  const instance = new BaiduPCS({
-    binary: execPath ?? binaryPath,
-    remotePath: remotePath ?? targetPath,
-  });
+  const { binary: binaryPath, target: targetPath } = getConfig(type);
+  const instance = createUploadInstance(type, execPath ?? binaryPath, remotePath ?? targetPath);
 
   const task = new SyncTask(
     instance,
@@ -52,7 +69,7 @@ export const loginByCookie = async ({
   cookie: string;
   execPath?: string;
 }) => {
-  const { binary: binaryPath } = getConfig();
+  const { binary: binaryPath } = getConfig("baiduPCS");
   const instance = new BaiduPCS({
     binary: execPath ?? binaryPath,
     remotePath: "",
@@ -60,11 +77,35 @@ export const loginByCookie = async ({
   return instance.loginByCookie(cookie);
 };
 
-export const isLogin = async ({ execPath }: { execPath?: string }) => {
-  const { binary: binaryPath } = getConfig();
-  const instance = new BaiduPCS({
-    binary: execPath ?? binaryPath,
-    remotePath: "",
-  });
+export const isLogin = async ({ execPath, type }: { execPath?: string; type: SyncType }) => {
+  const { binary: binaryPath } = getConfig(type);
+  const instance = createUploadInstance(type, execPath ?? binaryPath, "");
   return instance.isLoggedIn();
+};
+
+let aliyunpanLoginInstance: AliyunPan | null = null;
+
+export const aliyunpanLogin = async ({
+  execPath,
+  type,
+}: {
+  execPath?: string;
+  type: "getUrl" | "cancel" | "confirm";
+}) => {
+  const { binary: binaryPath } = getConfig("aliyunpan");
+  if (!aliyunpanLoginInstance) {
+    aliyunpanLoginInstance = new AliyunPan({
+      binary: execPath ?? binaryPath,
+      remotePath: "",
+    });
+  }
+  if (type === "getUrl") {
+    return aliyunpanLoginInstance.loginByBrowser();
+  } else if (type === "cancel") {
+    return aliyunpanLoginInstance.cancelLogin();
+  } else if (type === "confirm") {
+    return aliyunpanLoginInstance.confirmLogin();
+  } else {
+    throw new Error("Unsupported type");
+  }
 };